Cleanup run script for readability
